cart functionality
I removed alot of trial and error comments. The changes that are made allow a user to click to add an item to the cart, then inside the cart page, if they click the input to add more items it will add up the total cost of that row (not the total cost of all the products in the cart) and display it on the page.